+++ public/examples.me.js
@@ -3096,7 +3096,7 @@ var runThis = world => {
       matrixEngine.objLoader.initMeshBuffers(world.GL.gl, meshes[key]);
     }
     var textuteImageSamplers2 = {
-      source: ["res/images/armor.webp", "res/images/armor.webp"],
+      source: ["res/images/armor.webp"],
       mix_operation: "multiply"
     };
     var textuteImageSamplers = {
@@ -3109,19 +3109,13 @@ var runThis = world => {
     _manifest.default.scene.armor.LightsData.ambientLight.set(1, 1, 1);
     _manifest.default.scene.armor.position.z = -6;
     // App.scene.armor.mesh.setScale(3)
-
     setTimeout(() => {
       _manifest.default.scene.armor.activateShadows('spot');
       _manifest.default.scene.mac.activateShadows('spot');
-      // App.scene.armor.shadows.activeUpdate();
-      // App.scene.armor.shadows.animatePositionY();
     }, 2000);
-
-    // world.Add("cubeLightTex", 1, "MyCubeTex", textuteImageSamplers2);
-    // App.scene.MyCubeTex.activateShadows()
     world.Add("obj", 1, "mac", textuteImageSamplers, meshes.mac);
     _manifest.default.scene.mac.position.y = 1;
-    _manifest.default.scene.mac.position.x = -2;
+    _manifest.default.scene.mac.position.z = -6;
     _manifest.default.scene.mac.rotation.rotationSpeed.y = 20;
     _manifest.default.scene.mac.LightsData.ambientLight.set(1, 1, 1);
   }
@@ -3133,13 +3127,12 @@ var runThis = world => {
    * swap[0,2]
    * swap[1,3]
    * to switch x,y,z verts.
+   *  mac: "res/3d-objects/mac.obj"
    */
   matrixEngine.objLoader.downloadMeshes({
     armor: "res/3d-objects/armor.obj",
     mac: "res/3d-objects/mac.obj"
-  }, onLoadObj, {
-    swap: [0, 2]
-  });
+  }, onLoadObj);
 
   //delete images_local_var;
 };
@@ -6524,7 +6517,7 @@ var runThis = world => {
   world.Add("squareTex", 1, "MySquareTexure1", textuteImageSamplers);
   world.Add("cubeLightTex", 1, "MyCubeTexure1", textuteImageSamplers);
   App.scene.MyCubeTexure1.position.x = -2;
-
+  App.scene.MySquareTexure1.texParams.spriteAnimation.updateAfter = 1;
   // Access is App.scene.MySquareTexure1.texParams.spriteAnimation
   // this.shema = [4, 4]
   // this.updateAfter = 20;
@@ -9493,7 +9486,7 @@ _manifest.default.operation.triangle_buffer_procedure = function (object) {
   // console.log("Buffer the " + object.type + "'s color loaded success.");
 };
 _manifest.default.operation.obj_buffer_procedure = function (object) {
-  /* Vertex */
+  /* Vertex Color only - other come from objLoader */
   if (object.color && null !== object.shaderProgram.vertexColorAttribute) {
     object.vertexColorBuffer = _matrixWorld.world.GL.gl.createBuffer();
     _matrixWorld.world.GL.gl.bindBuffer(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.vertexColorBuffer);
@@ -9510,36 +9503,6 @@ _manifest.default.operation.obj_buffer_procedure = function (object) {
     object.vertexColorBuffer.itemSize = 4;
     object.vertexColorBuffer.numItems = 4;
   }
-
-  /* Texture
-    if (object.texture) {
-      // console.log("        Buffer the " + object.type + "'s texture");
-      object.vertexTexCoordBuffer = world.GL.gl.createBuffer();
-      world.GL.gl.bindBuffer(world.GL.gl.ARRAY_BUFFER, object.vertexTexCoordBuffer);
-      world.GL.gl.bufferData(world.GL.gl.ARRAY_BUFFER, object.geometry.texCoords , world.GL.gl.STATIC_DRAW);
-      object.vertexTexCoordBuffer.itemSize = 2;
-      object.vertexTexCoordBuffer.numItems = 24;
-    }
-  */
-
-  /* Normals                                   */
-  if (object.shaderProgram.useLightingUniform) {
-    // console.log("        Buffer the " + object.type + "'s normals");
-    // object.vertexNormalBuffer = world.GL.gl.createBuffer();
-    // world.GL.gl.bindBuffer(world.GL.gl.ARRAY_BUFFER, );
-    _matrixWorld.world.GL.gl.bufferData(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.normalBuffer, _matrixWorld.world.GL.gl.STATIC_DRAW);
-    // object.mesh.normalBuffer.itemSize = 3;
-    // object.mesh.normalBuffer.numItems = 24;
-  }
-
-  /* Indices
-    // console.log("        Buffer the " + object.type + "'s indices");
-    object.vertexIndexBuffer = world.GL.gl.createBuffer();
-    world.GL.gl.bindBuffer(world.GL.gl.ELEMENT_ARRAY_BUFFER, object.vertexIndexBuffer);
-    world.GL.gl.bufferData(world.GL.gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(object.geometry.indices), world.GL.gl.STATIC_DRAW);
-    object.vertexIndexBuffer.itemSize = 1;
-    object.vertexIndexBuffer.numItems = 36;
-  */
 };
 _manifest.default.operation.squareTex_buffer_procedure = function (object) {
   /* Vertex */
@@ -10256,7 +10219,7 @@ _manifest.default.operation.draws.cube = function (object, ray) {
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             }
-            if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+            if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
               //char * data; // data of 8-bit per sample RGBA image
               // int w, h; // size of the image
               // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -10331,7 +10294,7 @@ _manifest.default.operation.draws.cube = function (object, ray) {
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
           }
-          if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+          if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
             //char * data; // data of 8-bit per sample RGBA image
             // int w, h; // size of the image
             // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -10452,7 +10415,8 @@ _manifest.default.operation.draws.cube = function (object, ray) {
 
   // world.disableUnusedAttr(world.GL.gl, localLooper);
   // FIX FOR MOBILE - from white cube to normal texture view , good
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, 4);
+  // world.disableUnusedAttr(world.GL.gl, 4);
+
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
@@ -10637,8 +10601,10 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
     } else {
       // now to render the mesh test
       _matrixWorld.world.GL.gl.bindBuffer(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.vertexBuffer);
-      _matrixWorld.world.GL.gl.bufferData(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.vertices, _matrixWorld.world.GL.gl.STATIC_DRAW);
+      _matrixWorld.world.GL.gl.bufferData(_matrixWorld.world.GL.gl.ARRAY_BUFFER, new Float32Array(object.mesh.vertices), _matrixWorld.world.GL.gl.STATIC_DRAW);
       _matrixWorld.world.GL.gl.vertexAttribPointer(object.shaderProgram.vertexPositionAttribute, object.mesh.vertexBuffer.itemSize, _matrixWorld.world.GL.gl.FLOAT, false, 0, 0);
+      _matrixWorld.world.GL.gl.enableVertexAttribArray(object.shaderProgram.vertexPositionAttribute);
+      // console.log(".vertices.." + world.GL.gl.getBufferParameter(world.GL.gl.ARRAY_BUFFER, world.GL.gl.BUFFER_SIZE))
     }
   }
 
@@ -10711,10 +10677,9 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
     // disabled, then it needs to be re-enabled
     if (object.texture) {
       _matrixWorld.world.GL.gl.bindBuffer(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.textureBuffer);
-      _matrixWorld.world.GL.gl.enableVertexAttribArray(object.shaderProgram.textureCoordAttribute);
       _matrixWorld.world.GL.gl.vertexAttribPointer(object.shaderProgram.textureCoordAttribute, object.mesh.textureBuffer.itemSize, _matrixWorld.world.GL.gl.FLOAT, false, 0, 0);
-      //ori world.GL.gl.activeTexture(world.GL.gl.TEXTURE0);
-      // ori world.GL.gl.bindTexture(world.GL.gl.TEXTURE_2D, object.texture);
+      _matrixWorld.world.GL.gl.enableVertexAttribArray(object.shaderProgram.textureCoordAttribute);
+      // console.log(".tex.." + world.GL.gl.getBufferParameter(world.GL.gl.ARRAY_BUFFER, world.GL.gl.BUFFER_SIZE))
 
       if (object.streamTextures != null) {
         // video/webcam tex
@@ -10764,7 +10729,7 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
                 _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
                 _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               }
-              if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+              if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
                 //char * data; // data of 8-bit per sample RGBA image
                 // int w, h; // size of the image
                 // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -10833,7 +10798,7 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             }
-            if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+            if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
               //char * data; // data of 8-bit per sample RGBA image
               // int w, h; // size of the image
               // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -10865,11 +10830,6 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
       // world.GL.gl.disableVertexAttribArray(object.shaderProgram.textureCoordAttribute);
     }
   }
-
-  // Normals normalBuffer
-  //  world.GL.gl.bindBuffer(world.GL.gl.ARRAY_BUFFER, object.mesh.normalBuffer);
-  //  world.GL.gl.vertexAttribPointer(object.shaderProgram.vertexNormalAttribute, object.mesh.normalBuffer.itemSize, world.GL.gl.FLOAT, false, 0, 0);
-
   if (object.mesh.normalBuffer && object.shaderProgram.nMatrixUniform) {
     var normalMatrix = mat3.create();
     mat3.normalFromMat4(normalMatrix, object.mvMatrix);
@@ -10913,7 +10873,6 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
   }
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
-      // world.GL.gl.disable(world.GL.gl.DEPTH_TEST);
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
     }
     _matrixWorld.world.GL.gl.blendFunc(_matrixWorld.world.GL.gl[object.glBlend.blendParamSrc], _matrixWorld.world.GL.gl[object.glBlend.blendParamDest]);
@@ -10922,10 +10881,7 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
     _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.DEPTH_TEST);
   }
 
-  // ori this.setMatrixUniforms(object, this.pMatrix, object.mvMatrix);
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, localLooper + 1);
-  // world.GL.gl.drawElements(world.GL.gl[object.glDrawElements.mode], object.glDrawElements.numberOfIndicesRender, world.GL.gl.UNSIGNED_SHORT, 0);
-  // update for anim
+  // world.disableUnusedAttr(world.GL.gl, 4);
   _matrixWorld.world.GL.gl.drawElements(_matrixWorld.world.GL.gl[object.glDrawElements.mode], object.glDrawElements.numberOfIndicesRender, _matrixWorld.world.GL.gl.UNSIGNED_SHORT, 0);
   object.instancedDraws.overrideDrawArraysInstance(object);
   this.mvPopMatrix(object.mvMatrix, this.mvMatrixStack);
@@ -11081,7 +11037,7 @@ _manifest.default.operation.draws.drawSquareTex = function (object, ray) {
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             }
-            if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+            if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
               //char * data; // data of 8-bit per sample RGBA image
               // int w, h; // size of the image
               // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -11155,7 +11111,7 @@ _manifest.default.operation.draws.drawSquareTex = function (object, ray) {
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
           }
-          if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+          if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
             //char * data; // data of 8-bit per sample RGBA image
             // int w, h; // size of the image
             // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -11194,7 +11150,8 @@ _manifest.default.operation.draws.drawSquareTex = function (object, ray) {
   }
 
   // world.disableUnusedAttr( world.GL.gl, localLooper);
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, 4);
+  // world.disableUnusedAttr(world.GL.gl, 4);
+
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
@@ -11374,8 +11331,8 @@ _manifest.default.operation.draws.sphere = function (object, ray) {
     console.warn('WTF - ERROR10001');
   }
 
-  // world.disableUnusedAttr(world.GL.gl, localLooper);
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, 4);
+  // world.disableUnusedAttr(world.GL.gl, 4);
+
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
@@ -16683,6 +16640,8 @@ function defineworld(canvas, renderType) {
   world.initShaders = _engine.initShaders;
   world.handleLoadedTexture = _manifest.default.tools.BasicTextures;
   world.initTexture = _manifest.default.tools.loadTextureImage;
+
+  // DEPLACED
   world.disableUnusedAttr = function (gl, vertLimit) {
     var Local_looper1 = vertLimit;
     // var Local_looper1 = 0;
@@ -16691,6 +16650,9 @@ function defineworld(canvas, renderType) {
       gl.disableVertexAttribArray(Local_looper1);
       Local_looper1 = Local_looper1 + 1;
     }
+  };
+  world.disableAttr = function (gl, current) {
+    gl.disableVertexAttribArray(current);
   };
 
   // Bind base methods
@@ -17488,11 +17450,13 @@ function defineworld(canvas, renderType) {
               spriteAnimation: false,
               MIPMAP: false,
               ANISOTROPY: false,
-              TEXTURE_WRAP_S: this.GL.gl.REPEAT,
-              TEXTURE_WRAP_T: this.GL.gl.REPEAT,
+              TEXTURE_WRAP_S: null,
+              //this.GL.gl.REPEAT,
+              TEXTURE_WRAP_T: null,
+              // this.GL.gl.REPEAT,
               TEXTURE_MAG_FILTER: this.GL.gl.LINEAR,
               // NEAREST
-              TEXTURE_MIN_FILTER: this.GL.gl.LINEAR
+              TEXTURE_MIN_FILTER: this.GL.gl.LINEAR // LINEAR
             };
           } else {
             objObject.texParams = texturesPaths.params;
@@ -27559,9 +27523,9 @@ class MatrixSounds {
   }
   play(name) {
     if (this.audios[name].paused == true) {
-      this.audios[name].play();
+      return this.audios[name].play();
     } else {
-      this.tryClone(name);
+      return this.tryClone(name);
     }
   }
   tryClone(name) {
@@ -27570,7 +27534,9 @@ class MatrixSounds {
       while (this.audios[name + cc].paused == false) {
         cc++;
       }
-      if (this.audios[name + cc]) this.audios[name + cc].play();
+      if (this.audios[name + cc]) {
+        return this.audios[name + cc].play();
+      }
     } catch (err) {}
   }
 }