+++ public/me.lib.js
@@ -2617,7 +2617,7 @@ _manifest.default.operation.triangle_buffer_procedure = function (object) {
   // console.log("Buffer the " + object.type + "'s color loaded success.");
 };
 _manifest.default.operation.obj_buffer_procedure = function (object) {
-  /* Vertex */
+  /* Vertex Color only - other come from objLoader */
   if (object.color && null !== object.shaderProgram.vertexColorAttribute) {
     object.vertexColorBuffer = _matrixWorld.world.GL.gl.createBuffer();
     _matrixWorld.world.GL.gl.bindBuffer(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.vertexColorBuffer);
@@ -2634,36 +2634,6 @@ _manifest.default.operation.obj_buffer_procedure = function (object) {
     object.vertexColorBuffer.itemSize = 4;
     object.vertexColorBuffer.numItems = 4;
   }
-
-  /* Texture
-    if (object.texture) {
-      // console.log("        Buffer the " + object.type + "'s texture");
-      object.vertexTexCoordBuffer = world.GL.gl.createBuffer();
-      world.GL.gl.bindBuffer(world.GL.gl.ARRAY_BUFFER, object.vertexTexCoordBuffer);
-      world.GL.gl.bufferData(world.GL.gl.ARRAY_BUFFER, object.geometry.texCoords , world.GL.gl.STATIC_DRAW);
-      object.vertexTexCoordBuffer.itemSize = 2;
-      object.vertexTexCoordBuffer.numItems = 24;
-    }
-  */
-
-  /* Normals                                   */
-  if (object.shaderProgram.useLightingUniform) {
-    // console.log("        Buffer the " + object.type + "'s normals");
-    // object.vertexNormalBuffer = world.GL.gl.createBuffer();
-    // world.GL.gl.bindBuffer(world.GL.gl.ARRAY_BUFFER, );
-    _matrixWorld.world.GL.gl.bufferData(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.normalBuffer, _matrixWorld.world.GL.gl.STATIC_DRAW);
-    // object.mesh.normalBuffer.itemSize = 3;
-    // object.mesh.normalBuffer.numItems = 24;
-  }
-
-  /* Indices
-    // console.log("        Buffer the " + object.type + "'s indices");
-    object.vertexIndexBuffer = world.GL.gl.createBuffer();
-    world.GL.gl.bindBuffer(world.GL.gl.ELEMENT_ARRAY_BUFFER, object.vertexIndexBuffer);
-    world.GL.gl.bufferData(world.GL.gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(object.geometry.indices), world.GL.gl.STATIC_DRAW);
-    object.vertexIndexBuffer.itemSize = 1;
-    object.vertexIndexBuffer.numItems = 36;
-  */
 };
 _manifest.default.operation.squareTex_buffer_procedure = function (object) {
   /* Vertex */
@@ -3380,7 +3350,7 @@ _manifest.default.operation.draws.cube = function (object, ray) {
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             }
-            if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+            if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
               //char * data; // data of 8-bit per sample RGBA image
               // int w, h; // size of the image
               // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -3455,7 +3425,7 @@ _manifest.default.operation.draws.cube = function (object, ray) {
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
           }
-          if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+          if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
             //char * data; // data of 8-bit per sample RGBA image
             // int w, h; // size of the image
             // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -3576,7 +3546,8 @@ _manifest.default.operation.draws.cube = function (object, ray) {
 
   // world.disableUnusedAttr(world.GL.gl, localLooper);
   // FIX FOR MOBILE - from white cube to normal texture view , good
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, 4);
+  // world.disableUnusedAttr(world.GL.gl, 4);
+
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
@@ -3761,8 +3732,10 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
     } else {
       // now to render the mesh test
       _matrixWorld.world.GL.gl.bindBuffer(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.vertexBuffer);
-      _matrixWorld.world.GL.gl.bufferData(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.vertices, _matrixWorld.world.GL.gl.STATIC_DRAW);
+      _matrixWorld.world.GL.gl.bufferData(_matrixWorld.world.GL.gl.ARRAY_BUFFER, new Float32Array(object.mesh.vertices), _matrixWorld.world.GL.gl.STATIC_DRAW);
       _matrixWorld.world.GL.gl.vertexAttribPointer(object.shaderProgram.vertexPositionAttribute, object.mesh.vertexBuffer.itemSize, _matrixWorld.world.GL.gl.FLOAT, false, 0, 0);
+      _matrixWorld.world.GL.gl.enableVertexAttribArray(object.shaderProgram.vertexPositionAttribute);
+      // console.log(".vertices.." + world.GL.gl.getBufferParameter(world.GL.gl.ARRAY_BUFFER, world.GL.gl.BUFFER_SIZE))
     }
   }
 
@@ -3835,10 +3808,9 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
     // disabled, then it needs to be re-enabled
     if (object.texture) {
       _matrixWorld.world.GL.gl.bindBuffer(_matrixWorld.world.GL.gl.ARRAY_BUFFER, object.mesh.textureBuffer);
-      _matrixWorld.world.GL.gl.enableVertexAttribArray(object.shaderProgram.textureCoordAttribute);
       _matrixWorld.world.GL.gl.vertexAttribPointer(object.shaderProgram.textureCoordAttribute, object.mesh.textureBuffer.itemSize, _matrixWorld.world.GL.gl.FLOAT, false, 0, 0);
-      //ori world.GL.gl.activeTexture(world.GL.gl.TEXTURE0);
-      // ori world.GL.gl.bindTexture(world.GL.gl.TEXTURE_2D, object.texture);
+      _matrixWorld.world.GL.gl.enableVertexAttribArray(object.shaderProgram.textureCoordAttribute);
+      // console.log(".tex.." + world.GL.gl.getBufferParameter(world.GL.gl.ARRAY_BUFFER, world.GL.gl.BUFFER_SIZE))
 
       if (object.streamTextures != null) {
         // video/webcam tex
@@ -3888,7 +3860,7 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
                 _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
                 _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               }
-              if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+              if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
                 //char * data; // data of 8-bit per sample RGBA image
                 // int w, h; // size of the image
                 // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -3957,7 +3929,7 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             }
-            if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+            if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
               //char * data; // data of 8-bit per sample RGBA image
               // int w, h; // size of the image
               // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -3989,11 +3961,6 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
       // world.GL.gl.disableVertexAttribArray(object.shaderProgram.textureCoordAttribute);
     }
   }
-
-  // Normals normalBuffer
-  //  world.GL.gl.bindBuffer(world.GL.gl.ARRAY_BUFFER, object.mesh.normalBuffer);
-  //  world.GL.gl.vertexAttribPointer(object.shaderProgram.vertexNormalAttribute, object.mesh.normalBuffer.itemSize, world.GL.gl.FLOAT, false, 0, 0);
-
   if (object.mesh.normalBuffer && object.shaderProgram.nMatrixUniform) {
     var normalMatrix = mat3.create();
     mat3.normalFromMat4(normalMatrix, object.mvMatrix);
@@ -4037,7 +4004,6 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
   }
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
-      // world.GL.gl.disable(world.GL.gl.DEPTH_TEST);
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
     }
     _matrixWorld.world.GL.gl.blendFunc(_matrixWorld.world.GL.gl[object.glBlend.blendParamSrc], _matrixWorld.world.GL.gl[object.glBlend.blendParamDest]);
@@ -4046,10 +4012,7 @@ _manifest.default.operation.draws.drawObj = function (object, ray) {
     _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.DEPTH_TEST);
   }
 
-  // ori this.setMatrixUniforms(object, this.pMatrix, object.mvMatrix);
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, localLooper + 1);
-  // world.GL.gl.drawElements(world.GL.gl[object.glDrawElements.mode], object.glDrawElements.numberOfIndicesRender, world.GL.gl.UNSIGNED_SHORT, 0);
-  // update for anim
+  // world.disableUnusedAttr(world.GL.gl, 4);
   _matrixWorld.world.GL.gl.drawElements(_matrixWorld.world.GL.gl[object.glDrawElements.mode], object.glDrawElements.numberOfIndicesRender, _matrixWorld.world.GL.gl.UNSIGNED_SHORT, 0);
   object.instancedDraws.overrideDrawArraysInstance(object);
   this.mvPopMatrix(object.mvMatrix, this.mvMatrixStack);
@@ -4205,7 +4168,7 @@ _manifest.default.operation.draws.drawSquareTex = function (object, ray) {
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
               _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             }
-            if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+            if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
               //char * data; // data of 8-bit per sample RGBA image
               // int w, h; // size of the image
               // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -4279,7 +4242,7 @@ _manifest.default.operation.draws.drawSquareTex = function (object, ray) {
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MAG_FILTER, object.texParams.TEXTURE_MAG_FILTER | _matrixWorld.world.GL.gl.LINEAR);
             _matrixWorld.world.GL.gl.texParameteri(_matrixWorld.world.GL.gl.TEXTURE_2D, _matrixWorld.world.GL.gl.TEXTURE_MIN_FILTER, object.texParams.TEXTURE_MIN_FILTER | _matrixWorld.world.GL.gl.LINEAR);
           }
-          if (object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
+          if (typeof object.texParams.spriteAnimation != 'undefined' && object.texParams.spriteAnimation != false && object.textures[t].image.width != 0) {
             //char * data; // data of 8-bit per sample RGBA image
             // int w, h; // size of the image
             // int sx, sy, sw, sh; // position and size of area you want to crop.
@@ -4318,7 +4281,8 @@ _manifest.default.operation.draws.drawSquareTex = function (object, ray) {
   }
 
   // world.disableUnusedAttr( world.GL.gl, localLooper);
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, 4);
+  // world.disableUnusedAttr(world.GL.gl, 4);
+
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
@@ -4498,8 +4462,8 @@ _manifest.default.operation.draws.sphere = function (object, ray) {
     console.warn('WTF - ERROR10001');
   }
 
-  // world.disableUnusedAttr(world.GL.gl, localLooper);
-  _matrixWorld.world.disableUnusedAttr(_matrixWorld.world.GL.gl, 4);
+  // world.disableUnusedAttr(world.GL.gl, 4);
+
   if (object.glBlend.blendEnabled == true) {
     if (!_matrixWorld.world.GL.gl.isEnabled(_matrixWorld.world.GL.gl.BLEND)) {
       _matrixWorld.world.GL.gl.enable(_matrixWorld.world.GL.gl.BLEND);
@@ -9807,6 +9771,8 @@ function defineworld(canvas, renderType) {
   world.initShaders = _engine.initShaders;
   world.handleLoadedTexture = _manifest.default.tools.BasicTextures;
   world.initTexture = _manifest.default.tools.loadTextureImage;
+
+  // DEPLACED
   world.disableUnusedAttr = function (gl, vertLimit) {
     var Local_looper1 = vertLimit;
     // var Local_looper1 = 0;
@@ -9815,6 +9781,9 @@ function defineworld(canvas, renderType) {
       gl.disableVertexAttribArray(Local_looper1);
       Local_looper1 = Local_looper1 + 1;
     }
+  };
+  world.disableAttr = function (gl, current) {
+    gl.disableVertexAttribArray(current);
   };
 
   // Bind base methods
@@ -10612,11 +10581,13 @@ function defineworld(canvas, renderType) {
               spriteAnimation: false,
               MIPMAP: false,
               ANISOTROPY: false,
-              TEXTURE_WRAP_S: this.GL.gl.REPEAT,
-              TEXTURE_WRAP_T: this.GL.gl.REPEAT,
+              TEXTURE_WRAP_S: null,
+              //this.GL.gl.REPEAT,
+              TEXTURE_WRAP_T: null,
+              // this.GL.gl.REPEAT,
               TEXTURE_MAG_FILTER: this.GL.gl.LINEAR,
               // NEAREST
-              TEXTURE_MIN_FILTER: this.GL.gl.LINEAR
+              TEXTURE_MIN_FILTER: this.GL.gl.LINEAR // LINEAR
             };
           } else {
             objObject.texParams = texturesPaths.params;
@@ -20683,9 +20654,9 @@ class MatrixSounds {
   }
   play(name) {
     if (this.audios[name].paused == true) {
-      this.audios[name].play();
+      return this.audios[name].play();
     } else {
-      this.tryClone(name);
+      return this.tryClone(name);
     }
   }
   tryClone(name) {
@@ -20694,7 +20665,9 @@ class MatrixSounds {
       while (this.audios[name + cc].paused == false) {
         cc++;
       }
-      if (this.audios[name + cc]) this.audios[name + cc].play();
+      if (this.audios[name + cc]) {
+        return this.audios[name + cc].play();
+      }
     } catch (err) {}
   }
 }